Add a colour renderer
