+++ src/components/FeaturePanel/Properties/renderValue.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { getUrlForTag } from './getUrlForTag';
 import { slashToOptionalBr } from '../../helpers';
+import { osmColorToHex, whiteOrBlackText } from '../helpers/color';
+import styled from '@emotion/styled';
 
 const getEllipsisHumanUrl = (humanUrl) => {
   const MAX_LENGTH = 40;
@@ -39,12 +41,30 @@ const getHumanValue = (k, v, featured: boolean) => {
     return '✗';
   }
 
+  if (k === 'colour' && !v.startsWith('#')) {
+    // undocumented values result in no hex code
+    const hex = osmColorToHex(v);
+    return hex.startsWith('#') ? `${v} (${hex})` : v;
+  }
+
   return humanValue;
 };
 
-export const renderValue = (k, v, featured = false) => {
+const ColorValue = styled.div<{ v: string }>`
+  background-color: ${({ v }) => osmColorToHex(v)};
+  color: ${({ v }) => whiteOrBlackText(osmColorToHex(v))};
+  padding: 0.2rem 0.4rem;
+  border-radius: 0.125rem;
+  display: inline;
+`;
+
+export const renderValue = (k: string, v: string, featured = false) => {
   const url = getUrlForTag(k, v);
   const humanValue = getHumanValue(k, v, featured);
+
+  if (k === 'colour') {
+    return <ColorValue v={v}>{humanValue}</ColorValue>;
+  }
 
   return url ? <a href={url}>{slashToOptionalBr(humanValue)}</a> : humanValue;
 };
